requirements.txt
# =============================
# Core LangChain & LLM Frameworks
# =============================
langchain==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-huggingface==0.1.2

# =============================
# Hugging Face & Transformers
# =============================
transformers[sentencepiece]==4.46.3
huggingface_hub==0.26.2
sentencepiece==0.2.1
accelerate==1.1.1
bitsandbytes==0.44.1
protobuf>=4.25.0

# =============================
# Google Gemini & API Clients
# =============================
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-python-client==2.186.0

# =============================
# ML & Data Processing
# =============================
torch>=2.3.0
numpy>=1.26.4
pandas>=2.2.2
tqdm>=4.67.1

# =============================
# UI Framework
# =============================
streamlit==1.42.0

# =============================
# Utilities & Environment
# =============================
python-dotenv==1.0.1
filetype==1.2.0
requests>=2.32.3
pydantic>=2.10.9

# =============================
# Developer Extras (Optional)
# =============================
colorama>=0.4.6
tenacity>=9.1.2

